requirements.txt
Django==2.2
djangorestframework==3.9.2
pkg-resources==0.0.0
pytz==2025.2
sqlparse==0.4.4
